SExpr: add representation of "s-expression"
Add a definition of a s-expression.  The symbolic expression is defined
as:

```
type SExpr =
  | atom of Atom
  | list of SExpr list
```

An atom is either a string or numeric value.  In order to handle the
typing, we type the atom value.  As such, we define an atom as:

```
type Atom =
  | int of Int
  | double of Float
  | string of String
```

This allows for a definition of a typed s-expression that can be used
to define the s-expression processing for LISP.