SExpr: add support to lex a S-Expression
Add a simplistic tokeniser for an s-expression.  We simply munch a
series of '(', followed by any atoms, and finally munch any ')'.
Although we process them in order, because we lex it as:

```
(* [^\s()]* )*
```

we will process any invalid input as well.  The intent is that any
malformed stream will be detected at parse time.

TODO: Ideally we would track the source location for the input to
diagnose invalid input.