SExpr: support parsing of a token stream
Add a basic parser for a token stream.  This processes a token stream
into an AST representation.  The AST representation simply forms a tree
of s-expressions.  Add a set of interesting test cases for the parsing,
including the processing of a top level apply without an explicit
s-expression construction.

This allows us to process:

```
(lambda (x) x) 1
```

as if the user had written

```
((lambda (x) x) 1)
```