SExpr: add evaluation support
Add the definition of `Environment` which is the evaluation context.  It
is a mapping from an identifier (String) to a value.  The value may be
either a symbol or a procedure.

```
type Value =
  | value of Symbol
  | procedure of 'a -> Environment -> 'a
```

The procedure takes the environment as an `inout` parameter, allowing
mutation of the environment.  The default environment receives a number
of procedures by default, which are the builtin operations.  The initial
set of builtin operations are only sufficient for testing.  These
include:

- lambda: takes a list of free variables to bind, and a s-expr to
  evaluate in that context
- define: takes a list of a name and free variables to bind, and a
  s-expression to evaluate.
- quote: takes a list of atoms which is returned un-evaluated.

The evaluation process simply takes a given s-expression, and if it is
an atom, returning the value if it is a bound variable in the given
environment or itself if it is a string or the literal numeric constant
otherwise.

If the s-expression is a list, the count of the list determines the
behaviour.  If the list is 0 (`nil`), `nil` is returned.  If it has a
single element, the element is evaluated recursively and returned.
In any other case, if the first element is bound to a procedure, the
state of the procedure determines the behaviour.  If the bound procedure
is _not_ a special form procedure, the subsequent arguments are
evaluated eagerly.  Subsequently, the body of the bound procedure is
invoked with the arguments.

The returned value of the evaluation depends on the result.  If the
result is `nil`, the value is returned as is.  If the value is a list of
a single element, it is unwrapped to return a value.  In any other case,
the resulting list is re-evaluated in the new environment.

The result of an evaluation is always a single s-expression, which is
either an atom or `nil`.