SExpr: adjust `define` to handle variables
This adjust the builtin `define` to extend support for both functions
and values.  The builtin as defined as:

`define` `binding` `body`

If the binding is a tuple, the first name is taken to be the name and
all other are free variables which are bound.  If the binding is a
single atom, the expression is evaluated as a variable.